Finished rest

+++ spring-jpa/spring-jpa/src/main/java/com/example/spring_jpa/service/UserService.java
@@ -2,18 +2,21 @@ package com.example.spring_jpa.service;
 
 import com.example.spring_jpa.repository.UserRepository;
 import com.example.spring_jpa.model.User;
+import com.example.spring_jpa.exception.ResourceNotFoundException;
+//import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class UserService {
-
+    
     private final UserRepository userRepository;
 
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
+
 
     public List<User> findAllUsers() {
         return userRepository.findAll();
@@ -24,11 +27,23 @@ public class UserService {
     }
 
     public Optional<User> getUserById(Long id) {
-        return userRepository.findById(id);
+        try{
+             return userRepository.findById(id);
+        } catch (Exception e) {
+            throw new ResourceNotFoundException("User not found with id: " + id);
+        }
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        try{
+            if (userRepository.findById(id).isEmpty()) {
+                throw new ResourceNotFoundException("User not found with id: " + id);
+            } else {
+                userRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            throw new ResourceNotFoundException("User not found with id: " + id);
+        }
     }
 
     public User getUserByEmail(String email) {
@@ -36,6 +51,7 @@ public class UserService {
     }
 
     public Optional<User> updateUser(User user) {
+        System.out.println("Updating user with ID: " + user.getId());
         User existingUser = userRepository.findById(user.getId()).orElse(null);
         if (existingUser != null) {
             existingUser.setFirstName(user.getFirstName());
@@ -48,8 +64,8 @@ public class UserService {
         return Optional.empty();
     }
 
-    public List<User> findTop10Users() {
-        return userRepository.findFirst10ByOrderByUsernameAsc();
+    public List<User> findTop10Users(String username) {
+         return userRepository.findTop10ByUsernameContainingIgnoreCaseOrderByUsernameAsc(username);
     }
 
     public Long countUsers() {
